+++ sdk/eventhubs/azure-messaging-eventhubs/test/eventhubs-stress-test/eventhubs_stress_test.cpp
@@ -117,7 +117,7 @@ private:
       event.Properties["PartitionId"]
           = static_cast<Azure::Core::Amqp::Models::AmqpValue>(m_partitionId);
       AddEndProperty(event, m_numberToSend);
-      batch.AddMessage(event);
+      batch.TryAddMessage(event);
     }
     m_client->Send(batch, context);