+++ sdk/eventhubs/azure-messaging-eventhubs/test/ut/consumer_client_test.cpp
@@ -30,7 +30,7 @@ namespace Azure { namespace Messaging { namespace EventHubs { namespace Test {
 
       Azure::Messaging::EventHubs::ProducerClient producer{connStringNoEntityPath, eventHubName};
       EventDataBatch batch{producer.CreateBatch()};
-      EXPECT_TRUE(batch.AddMessage(Models::EventData{"Test"}));
+      EXPECT_TRUE(batch.TryAddMessage(Models::EventData{"Test"}));
       EXPECT_NO_THROW(producer.Send(batch));
     }
   };