+++ sdk/eventhubs/azure-messaging-eventhubs/test/ut/round_trip_test.cpp
@@ -30,7 +30,7 @@ namespace Azure { namespace Messaging { namespace EventHubs { namespace Test {
       Azure::Messaging::EventHubs::EventDataBatchOptions batchOptions;
       batchOptions.PartitionId = "1";
       Azure::Messaging::EventHubs::EventDataBatch eventBatch{producer.CreateBatch(batchOptions)};
-      eventBatch.AddMessage(Azure::Messaging::EventHubs::Models::EventData("Hello world!"));
+      eventBatch.TryAddMessage(Azure::Messaging::EventHubs::Models::EventData("Hello world!"));
       EXPECT_NO_THROW(producer.Send(eventBatch));
     }
 
@@ -66,7 +66,7 @@ namespace Azure { namespace Messaging { namespace EventHubs { namespace Test {
       Azure::Messaging::EventHubs::EventDataBatchOptions batchOptions;
       batchOptions.PartitionId = "1";
       Azure::Messaging::EventHubs::EventDataBatch eventBatch{producer.CreateBatch(batchOptions)};
-      eventBatch.AddMessage(Azure::Messaging::EventHubs::Models::EventData({1, 2, 3, 4, 5}));
+      eventBatch.TryAddMessage(Azure::Messaging::EventHubs::Models::EventData({1, 2, 3, 4, 5}));
       EXPECT_NO_THROW(producer.Send(eventBatch));
     }
 
@@ -119,7 +119,7 @@ namespace Azure { namespace Messaging { namespace EventHubs { namespace Test {
       eventData.Body = {1, 2, 3, 4, 5, 6, 7};
       eventData.ContentType = "application/binary";
       eventData.MessageId = "Test Message Id";
-      EXPECT_TRUE(eventBatch.AddMessage(eventData));
+      EXPECT_TRUE(eventBatch.TryAddMessage(eventData));
       EXPECT_NO_THROW(producer.Send(eventBatch));
     }